# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "TodoList 项目API文档",
        "title": "TodoList 项目API文档",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "jin quan",
            "email": "bothwin.qj@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "这里写接口服务的host",
    "basePath": "/",
    "paths": {
        "/user/todo_list": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "TODOList"
                ],
                "summary": "获取用户待办列表接口",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "user_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/main.TodoResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.ToDoListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/main.TodoInfoResponse"
                    }
                },
                "user_name": {
                    "type": "string"
                }
            }
        },
        "main.TodoInfoResponse": {
            "type": "object",
            "properties": {
                "description": {
                    "description": "描述",
                    "type": "string"
                },
                "id": {
                    "description": "ID",
                    "type": "integer"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                },
                "score": {
                    "description": "分数 1-100",
                    "type": "integer"
                }
            }
        },
        "main.TodoResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "response": {
                    "$ref": "#/definitions/main.ToDoListResponse"
                }
            }
        }
    }
}